vinyl: throttle tx thread if dump doesn't catch up
If tx_write_rate > dump_bandwidth/write_amp, in the time it takes the tx
thread to insert X bytes, the dump thread will only be able to write

  Y = dump_bandwidth/write_amp * X / tx_write_rate  bytes

(here write_amp is write amplification)

So to let the dump thread catch up, we should throttle the tx thread
for

  (X - Y) / (dump_bandwidth/write_amp)  seconds

which is equal to

  X * (1 / (dump_bandwidth/write_amp) - 1 / tx_write_rate)

To avoid wasting cpu cycles and piling up fibers, we do not use
fiber_sleep() for throttling transactions. Instead we simply tune
io_collect_interval of the tx event loop.