vinyl: reserve quota before using memory
Currently, we first allocate memory and only then use quota, throttling
the current transaction after it actually ate memory. As a result, the
quota limit doesn't hold in case of multiple clients. To fix that, let's
estimate the amount of memory the transaction is going to use and
reserve quota accordingly before submitting the transaction to the WAL,
i.e. in vy_prepare(), and fix the quota in vy_commit() in case the
transaction actually consumed more/less memory.