vinyl: use vy_task.run_count and vy_task.compact_from
When we will have single mem per index, there will be cases,
when distribution of the index runs and compaction of a range
are working concurrently.
Index 'distribute_complete' callback can update compact_priority
before ending of the range compaction, and vy_task_compact_complete()
will delete the new run due to the compact_priority.

So we must save the compact_priority in the compaction task sheduling
and save the run, from which to start to remove runs.