vinyl: separate discard and prepare of a run and the range.new_run
The ranges in index->tree will be supplemented by new runs from
'distribute' task. On the beginning of distribution we must
create a new run, not hardlinked with any range, because in the
end of distribution the run actually can be linked with another
range. And in addition, the range can be deleted during split_abort.